requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
discord.py==2.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
